chore: fix dependency arrangement
