+++ packages/fetchai/connections/http_server/tests/test_http_server.py
@@ -18,6 +18,9 @@
 #
 # ------------------------------------------------------------------------------
 """This module contains the tests of the HTTP Server connection module."""
+# type: ignore # noqa: E800
+# pylint: skip-file
+
 import asyncio
 import logging
 import os
