+++ plugins/aea-ledger-ethereum/aea_ledger_ethereum/test_tools/fixture_helpers.py
@@ -22,9 +22,7 @@
 from contextlib import contextmanager
 from pathlib import Path
 from typing import Dict
-from urllib.parse import urlparse
 
-import docker
 import pytest
 from aea_ledger_ethereum.test_tools.constants import (
     ETHEREUM_PRIVATE_KEY_PATH,
@@ -61,6 +59,8 @@ def _ganache_context(
     timeout: float = 2.0,
     max_attempts: int = 10,
 ):
+    import docker
+
     client = docker.from_env()
     image = GanacheDockerImage(
         client, ganache_addr, ganache_port, config=ganache_configuration
