+++ scripts/whitelist.py
@@ -307,3 +307,13 @@ to_v1_string  # unused function (aea/cli/ipfs_hash.py:382)
 BTC_ENCODING  # unused variable (aea/helpers/cid.py:32)
 is_cid  # unused method (aea/helpers/cid.py:200)
 launch_image  # unused function (aea/test_tools/docker_image.py:183)
+UNKNOWN_PROTOCOL_PUBLIC_ID  # unused variable (aea/test_tools/constants.py:28)
+UNKNOWN_CONNECTION_PUBLIC_ID  # unused variable (aea/test_tools/constants.py:29)
+_client  # unused attribute (aea/test_tools/docker_image.py:51)
+sleep_rate # unused variable (aea/test_tools/docker_image.py:107)
+AnyStringWith  # unused class (aea/test_tools/mocks.py:29)
+RegexComparator  # unused class (aea/test_tools/mocks.py:41)
+ctx_mock_Popen  # unused function (aea/test_tools/mocks.py:55)
+get_unused_tcp_port  # unused function (aea/test_tools/network.py:29)
+get_host  # unused function (aea/test_tools/network.py:39)
+wait_for_condition  # unused function (aea/test_tools/utils.py:26)
